requirements.txt
os-sys
wxpython
pycryptodome
piexif
pydrive

